zebra: fix if_zebra_speed_update timer
Currently, if_zebra_speed_update is set after interface creation.
This can create a race condition that links interface with ifindex 0 to ns, and causes the following problems:
1. BGP neighbor on that interface is permanently down
2. zebra asserts if more than one interfaces are affected
Fix this by moving the timer after if_add_update, and calls zebra_interface_add_update instead of if_add_update.
Also reschedule the timer if there is an error or the speed obtained is 0.

Signed-off-by: markxiao <markxiao@arista.com>